[libsailfishapp] Let Qt Creator enable QML debugging by default for QML-only apps. Contribute to JB#7833
This is NOT to enable the app to listen for QML debugger connection by
default (on device). This is to enable QML debugging in Qt Creator
project run settings by default. Otherwise it would have to be enabled
manually in order to start QML debugging from Qt Creator.

See QmakeProjectManager::QmakeProject::updateQmlJSCodeModel() in Qt
Creator sources.